Separacion funciones y creación de interfaz gráfica
